UHM-5994 - Initial commit to move all of the printer functionality in the mirebalais API project over to pihcore.  Both modules build and test and the system starts up fine after this change.
