--- api/src/main/java/org/openmrs/module/mirebalais/printer/HC100PrintHandler.java
@@ -1,19 +1,0 @@
-package org.openmrs.module.mirebalais.printer;
-
-import org.openmrs.module.printer.handler.SocketPrintHandler;
-
-public class HC100PrintHandler extends SocketPrintHandler {
-
-    @Override
-    public String getDisplayName() {
-        return "HC100 Print Handler";
-    }
-
-    @Override
-    public String getBeanName() {
-        return "hc100PrintHandler";
-    }
-
-    // just a simple socket print handlers
-
-}
