+++ api/src/main/java/org/openmrs/module/mirebalais/setup/ConfigurationSetup.java
@@ -48,6 +48,7 @@ import org.openmrs.module.pihcore.setup.MetadataSharingSetup;
 import org.openmrs.module.pihcore.setup.NameTemplateSetup;
 import org.openmrs.module.pihcore.setup.PacIntegrationSetup;
 import org.openmrs.module.pihcore.setup.PatientIdentifierSetup;
+import org.openmrs.module.pihcore.setup.PrinterSetup;
 import org.openmrs.module.pihcore.setup.ReportSetup;
 import org.openmrs.module.printer.PrinterService;
 import org.openmrs.module.registrationcore.RegistrationCoreConstants;
