+++ omod/src/main/java/org/openmrs/module/mirebalais/page/controller/patientRegistration/PrintIdCardPageController.java
@@ -4,8 +4,8 @@ import org.apache.commons.lang3.StringUtils;
 import org.openmrs.Location;
 import org.openmrs.Patient;
 import org.openmrs.module.appui.UiSessionContext;
-import org.openmrs.module.mirebalais.printer.IdPrinter;
-import org.openmrs.module.mirebalais.printer.impl.ZlEmrIdCardPrinter;
+import org.openmrs.module.pihcore.printer.IdPrinter;
+import org.openmrs.module.pihcore.printer.impl.ZlEmrIdCardPrinter;
 import org.openmrs.module.pihcore.config.Config;
 import org.openmrs.module.reporting.common.ObjectUtil;
 import org.openmrs.ui.framework.UiUtils;
