+++ pom.xml
@@ -588,91 +588,6 @@
             <scope>provided</scope>
         </dependency>
 
-        <!-- dependencies required for the Zebra ZXP Series 3 API -->
-
-        <dependency>
-            <groupId>net.sf.barcode4j</groupId>
-            <artifactId>barcode4j</artifactId>
-            <version>2.1</version>
-            <exclusions>
-                <exclusion>
-                    <artifactId>commons-lang</artifactId>
-                    <groupId>commons-lang</groupId>
-                </exclusion>
-                <exclusion>
-                    <groupId>commons-logging</groupId>
-                    <artifactId>commons-logging</artifactId>
-                </exclusion>
-                <exclusion>
-                    <groupId>commons-cli</groupId>
-                    <artifactId>commons-cli</artifactId>
-                </exclusion>
-            </exclusions>
-        </dependency>
-
-        <dependency>
-            <groupId>commons-net</groupId>
-            <artifactId>commons-net</artifactId>
-            <version>3.1</version>
-        </dependency>
-
-        <dependency>
-            <groupId>com.google.code.gson</groupId>
-            <artifactId>gson</artifactId>
-            <version>2.2</version>
-        </dependency>
-
-        <dependency>
-            <groupId>jaxen</groupId>
-            <artifactId>jaxen</artifactId>
-            <version>1.1.6</version>
-        </dependency>
-
-        <dependency>
-            <groupId>com.opencsv</groupId>
-            <artifactId>opencsv</artifactId>
-            <version>3.1</version>
-        </dependency>
-
-        <dependency>
-            <groupId>westhawk</groupId>
-            <artifactId>snmp</artifactId>
-            <version>4_13</version>
-        </dependency>
-
-        <dependency>
-            <groupId>org.snmp4j</groupId>
-            <artifactId>snmp4j</artifactId>
-            <version>1.10.1</version>
-            <exclusions>
-                <exclusion>
-                    <artifactId>log4j</artifactId>
-                    <groupId>log4j</groupId>
-                </exclusion>
-            </exclusions>
-        </dependency>
-
-        <dependency>
-            <groupId>com.google.zxing</groupId>
-            <artifactId>core</artifactId>
-            <version>1.7</version>
-        </dependency>
-
-        <!-- added manually to OpenMRS maven repo since this is not currently mavenized -->
-        <dependency>
-            <groupId>com.zebra.sdk</groupId>
-            <artifactId>ZSDK_API</artifactId>
-            <version>2.12.3782</version>
-        </dependency>
-
-        <dependency>
-            <groupId>com.zebra.sdk</groupId>
-            <artifactId>ZSDK_CARD_API</artifactId>
-            <version>2.12.3782</version>
-        </dependency>
-
-        <!-- end dependencies for Zebra ZXP Series 3 API -->
-
         <!-- required to fire up the test context, used by registration core -->
         <dependency>
             <groupId>com.fasterxml.jackson.core</groupId>
